pyproject.toml
[project]
name = "nonebot-plugin-steam-game-status"
version = "0.2.0"
description = "播报群友 Steam 游戏状态"
authors = [
    {name = "nek0us", email = "nekouss@mail.com"},
]
license = {text = "GPL3"}
dependencies = ["nonebot2>=2.0.0", "nonebot-plugin-alconna>=0.57.0", "nonebot-plugin-apscheduler>=0.2.0","nonebot_plugin_htmlrender>=0.3.1","nonebot_plugin_localstore"]
requires-python = ">=3.10"
readme = "README.md"

[project.urls]
Homepage = "https://github.com/nek0us/nonebot_plugin_steam_game_status"
Repository = "https://github.com/nek0us/nonebot_plugin_steam_game_status"

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"